#26 Move article class and detex function to article class
