+++ astro_ph/search.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 # standard library
 import asyncio
-from dataclasses import dataclass, replace
+from dataclasses import dataclass
 from datetime import datetime, timedelta
 from typing import AsyncIterable, Awaitable, Iterable, Optional, Sequence, Union
 
@@ -12,7 +12,7 @@ from typing import AsyncIterable, Awaitable, Iterable, Optional, Sequence, Union
 from aiohttp import ClientSession, TCPConnector
 from feedparser import FeedParserDict, parse
 from typing_extensions import Final
-from .detex import detex
+from .article import Article
 
 
 # constants
@@ -22,33 +22,9 @@ N_CONCURRENT: Final[int] = 1
 N_MAX_ARTICLES: Final[int] = 1000
 N_PER_REQUEST: Final[int] = 100
 SECOND: Final[timedelta] = timedelta(seconds=1)
-SEPARATOR: Final[str] = "++++++++++"
 
 
 # data classes
-@dataclass
-class Article:
-    """Article class for storing article information."""
-
-    title: str  #: Title of an article.
-    authors: Sequence[str]  #: Author(s) of an article.
-    summary: str  #: Summary of an article.
-    arxiv_url: str  #: arXiv URL of an article.
-
-    def __post_init__(self) -> None:
-        self.title = detex(self.title)
-        self.summary = detex(self.summary)
-
-    def replace(self, text: str, translated: str) -> Article:
-        """Method necessary to become translatable."""
-        title, summary = translated.split(SEPARATOR)
-        return replace(self, title=title, summary=summary)
-
-    def __str__(self) -> str:
-        """Method necessary to become translatable."""
-        return f"{self.title}\n{SEPARATOR}\n{self.summary}"
-
-
 @dataclass
 class Search:
     """Search class for searching for articles in arXiv."""
